# internal/service/resources/api/openapi.yaml
...
openapi: "3.0.0"
info:
  title: O2IMS Infrastructure Inventory API
  version: 1.0.0
  description: |
    API for O2IMS Infrastructure Inventory
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
- url: http://localhost:8000
...
paths:
  /o2ims-infrastructureInventory/api_versions:
    get:
      operationId: getAllVersions
      summary: Get API versions
      description: |
        Returns the complete list of API versions implemented by the service.
      tags:
      - metadata
      responses:
        '200':
          description: |
            Successfully obtained the complete list of versions.
          content:
            application/json:
              schema:
                $ref: "../../common/api/openapi.yaml#/components/schemas/APIVersions"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
...
  /o2ims-infrastructureInventory/v1/api_versions:
    get:
      operationId: getMinorVersions
      summary: Get minor API versions
      description: |
        Returns the list of minor API versions implemented for this major version of the API.
      tags:
      - metadata
      responses:
        '200':
          description: |
            Success
          content:
            application/json:
              schema:
                $ref: "../../common/api/openapi.yaml#/components/schemas/APIVersions"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
...
  /o2ims-infrastructureInventory/v1/subscriptions:
    get:
      operationId: getSubscriptions
      summary: Get subscriptions
      description: |
        Returns the list of inventory subscriptions.
      parameters:
      - $ref: "../../common/api/openapi.yaml#/components/parameters/excludeFields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/allFields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/fields"
      - $ref: "../../common/api/openapi.yaml#/components/parameters/filter"
      tags:
      - subscriptions
      responses:
        '200':
          description: |
            Successfully obtained the list of subscriptions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
    post:
      operationId: createSubscription
      summary: Create subscriptions
      description: |
        Creates an inventory subscriptions.
      tags:
      - subscriptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Subscription'
      responses:
        '201':
          description: |
            Successfully created the subscriptions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
  /o2ims-infrastructureInventory/v1/subscriptions/{subscriptionId}:
    get:
      operationId: getSubscription
      summary: Get subscription
      description: |
        Returns the details of a subscription.
      parameters:
      - $ref: "#/components/parameters/subscriptionId"
      tags:
      - subscriptions
      responses:
        '200':
          description: |
            Successfully obtained the details of the subscription.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        '400':
          description: Bad request
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '404':
          description: The specified entity was not found.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
    delete:
      operationId: deleteSubscription
      summary: Delete subscription
      description: |
        Deletes a subscription.
      parameters:
      - $ref: "#/components/parameters/subscriptionId"
      tags:
      - subscriptions
      responses:
        '200':
          description: |
            Successfully deleted the subscription.
        '404':
          description: The specified entity was not found.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '../../common/api/openapi.yaml#/components/schemas/ProblemDetails'
...
components:
  parameters:
...
    subscriptionId:
      name: subscriptionId
      description: |
        Unique identifier of a subscription.
      in: path
      required: true
      schema:
        type: string
        format: uuid
      example: aa83d0ba-a5ee-4f1f-be84-f334b21166cf
...
  schemas:
...
    Subscription:
      description: |
        Information about an inventory subscription.
      type: object
      properties:
        subscriptionId:
          type: string
          format: uuid
          readOnly: true
          description: Identifier for the Subscription. This identifier is allocated by the O-Cloud.
          example: "65221564-8b05-416a-bfc3-c250bc64e1aa"
        consumerSubscriptionId:
          type: string
          format: uuid
          description: Identifier for the consumer of events sent due to the Subscription.
        filter:
          type: string
          description: |
            Criteria for events which do not need to be reported or will be filtered by the subscription 
            notification service. Therefore, if a filter is not provided then all events are reported.
        callback:
          type: string
          description: |
            The fully qualified URI to a consumer procedure which can process a Post of the 
            InventoryEventNotification.
          example: https://smo.example.com/smo/v1/ocloud_inventory_observer
      required:
      - callback
...